separate read and write, move to different package
